modules/location_producer/requirements.txt
grpcio==1.34.0
grpcio-tools==1.34.0
kafka-python==2.0.2
protobuf